+++ CHANGES.rst
@@ -3,9 +3,14 @@ Changelog
 
 A list of changes between each release.
 
+0.3.0 (2018-12-14)
+^^^^^^^^^^^^^^^^^^
+- Switch device image back to avatar url (fixed)
+- Preserve last activity image functionality via `device.activity_image`
+
 0.2.0 (2018-12-07)
 ^^^^^^^^^^^^^^^^^^
-- Swtich device image from avatar url (broken) to last activity image.
+- Switch device image from avatar url (broken) to last activity image.
 
 0.1.0 (2017-09-24)
 ^^^^^^^^^^^^^^^^^^